refactor(precompiles): extract announce entity into its own module
Addresses PR review: the `AnnounceCall` abstraction, the borrowed representation
(`BorrowedAnnounce` + its decode), and the two trait impls were inline at the bottom
of `dispatch.rs`. They're a cohesive "announcement" data entity, so move them into a
new `b20_asset/announce.rs`, mirroring how `AssetCall` is declared and re-exported.

The borrowed decode moves onto the type it produces (`BorrowedAnnounce::decode`,
was `B20AssetToken::decode_announce_borrowed`). `run_announce` and the `route`
interception stay in `dispatch.rs` — the internal-call loop lives there deliberately
because re-dispatching sub-calls is a routing responsibility (per the module doc).

Pure structural move: no behavior, accept-set, error-byte, or gas change. All 708
crate tests pass unchanged (only the one direct-decode test call was renamed to
`BorrowedAnnounce::decode`).
